+++ kinto/core/resource/__init__.py
@@ -709,7 +709,6 @@ class UserResource(object):
             }
             response = http_error(HTTPNotFound(), errno=ERRORS.INVALID_RESOURCE_ID,
                                   details=details)
-
             raise response
 
     def _add_timestamp_header(self, response, timestamp=None):