+++ tests/test_views_records.py
@@ -38,7 +38,15 @@ class RecordsViewTest(BaseWebTest, unittest.TestCase):
 
     def test_unknown_collection_raises_404(self):
         other_collection = self.collection_url.replace('barley', 'pills')
-        self.app.get(other_collection, headers=self.headers, status=404)
+        resp = self.app.get(other_collection, headers=self.headers, status=404)
+        self.assertEqual(resp.json['details']['id'], 'pills')
+        self.assertEqual(resp.json['details']['resource_name'], 'collection')
+
+    def test_unknown_record_raises_404(self):
+        other_record = self.record_url.replace(self.record['id'], self.record['id']+'blah')
+        response = self.app.get(other_record, headers=self.headers, status=404)
+        self.assertEqual(response.json['details']['id'], self.record['id']+'blah')
+        self.assertEqual(response.json['details']['resource_name'], 'record')
 
     def test_unknown_collection_does_not_query_timestamp(self):
         other_collection = self.collection_url.replace('barley', 'pills')